WIP: Update unit tests
