+++ packages/evo-objects/tests/test_tables.py
@@ -10,18 +10,15 @@
 #  limitations under the License.
 
 import hashlib
-import random
 import unittest
 import uuid
 from collections.abc import Iterator
-from datetime import datetime, timezone
 from io import BytesIO
 from pathlib import Path
 from typing import BinaryIO
 from unittest import mock
 from uuid import UUID
 
-import numpy
 import pyarrow as pa
 import pyarrow.parquet as pq
 from parameterized import parameterized, parameterized_class
@@ -37,7 +34,7 @@ from evo.objects.utils import (
     all_known_formats,
     get_known_format,
 )
-from evo.objects.utils.tables import _ColumnFormat
+from helpers import get_sample_table
 
 SAMPLE_DATA_LENGTH = 10
 ENVIRONMENT = Environment(hub_url=BASE_URL, org_id=UUID(int=0), workspace_id=UUID(int=0))
@@ -64,101 +61,6 @@ def setUpModule() -> None:
 def _all_known_formats_for_testing() -> Iterator[dict]:
     for known_format in all_known_formats:
         yield {"data_format": known_format, "expected_field_names": known_format._field_names}
-
-
-def _generate_float64_data(n_samples: int) -> Iterator[float]:
-    max_ = numpy.finfo("float64").max
-    for _ in range(n_samples):
-        yield max_ * random.uniform(-1.0, 1.0)
-
-
-def _generate_int_data(int_type: str, n_samples: int) -> Iterator[int]:
-    min_, max_ = numpy.iinfo(int_type).min, numpy.iinfo(int_type).max
-    for _ in range(n_samples):
-        yield random.randint(min_, max_)
-
-
-def _generate_bool_data(n_samples: int) -> Iterator[bool]:
-    for _ in range(n_samples):
-        yield random.choice((True, False))
-
-
-def _generate_string_data(n_samples: int) -> Iterator[str]:
-    str_sample = "0123456789ABCDEF "
-    for _ in range(n_samples):
-        length = random.randint(10, 10000)
-        yield "".join(random.choices(str_sample, k=length))
-
-
-def _generate_timestamp_data(n_samples: int) -> Iterator[datetime]:
-    min_ = datetime(1970, 1, 1, tzinfo=timezone.utc).timestamp()
-    max_ = datetime(2038, 12, 31, 23, 59, 59, 999999, tzinfo=timezone.utc).timestamp()
-    for _ in range(n_samples):
-        yield datetime.utcfromtimestamp(random.uniform(min_, max_))
-
-
-def _generate_data(format_id: str, n_samples: int) -> Iterator:
-    match format_id:
-        case "float64":
-            yield from _generate_float64_data(n_samples)
-        case "uint8" | "uint32" | "uint64" | "int32" | "int64" as int_type:
-            yield from _generate_int_data(int_type, n_samples)
-        case "bool":
-            yield from _generate_bool_data(n_samples)
-        case "string":
-            yield from _generate_string_data(n_samples)
-        case "timestamp":
-            yield from _generate_timestamp_data(n_samples)
-        case unknown_format:
-            raise TypeError(f"Unsupported format '{unknown_format}'")
-
-
-def _get_sample_column(column_format: _ColumnFormat, n_samples: int) -> pa.Array:
-    return pa.array(_generate_data(column_format.id, n_samples), type=column_format.type, size=n_samples)
-
-
-def _get_table_schema(columns: list[_ColumnFormat]) -> pa.Schema:
-    return pa.schema([pa.field(f"{column.id}[{i}]", column.type, nullable=False) for i, column in enumerate(columns)])
-
-
-def _change_format(current_format: _ColumnFormat) -> _ColumnFormat:
-    match current_format.id:
-        case "float64":
-            return _ColumnFormat("int64")
-        case "uint8" | "uint32" | "uint64" | "int32" | "int64":
-            return _ColumnFormat("float64")
-        case "bool" | "timestamp":
-            return _ColumnFormat("string")
-        case "string":
-            return _ColumnFormat("bool")
-        case unknown_format:
-            raise TypeError(f"Unsupported format '{unknown_format}'")
-
-
-def _get_sample_table(
-    table_format: BaseTableFormat, n_rows: int, add_column: bool = False, change_types: bool = False
-) -> pa.Table:
-    column_formats = [column for column in table_format._columns]
-
-    if add_column:
-        column_formats.append(_ColumnFormat(column_formats[-1].type))
-
-    if change_types:
-        column_formats = [_change_format(column) for column in column_formats]
-
-    if table_format._multi_dimensional:
-        # Test multidimensional tables with an arbitrary number of columns. If the number of columns matches a more
-        # specific GO type (one with a fixed number of columns), the more specific type would be instantiated.
-        column_formats *= 20
-
-    sample_schema = pa.schema(
-        [pa.field(f"{column.id}[{i}]", column.type, nullable=False) for i, column in enumerate(column_formats)]
-    )
-    sample_data = [
-        pa.array(_generate_data(column_format.id, n_rows), type=column_format.type, size=n_rows)
-        for column_format in column_formats
-    ]
-    return pa.table(sample_data, names=sample_schema.names).cast(sample_schema)
 
 
 def _get_buffer_digest(buffer: BinaryIO) -> str:
@@ -189,7 +91,7 @@ class TestKnownFormat(unittest.TestCase):
     expected_field_names: str
 
     def setUp(self) -> None:
-        self.sample_table = _get_sample_table(table_format=self.data_format, n_rows=SAMPLE_DATA_LENGTH)
+        self.sample_table = get_sample_table(table_format=self.data_format, n_rows=SAMPLE_DATA_LENGTH)
         self.expected_parquet_digest = _get_table_digest(self.sample_table)
         self.data_dir = CACHE.get_location(ENVIRONMENT, self.__class__.__name__)
         self.parquet_file = self.data_dir / self.expected_parquet_digest
@@ -251,7 +153,7 @@ class TestKnownFormat(unittest.TestCase):
             sample_length = SAMPLE_DATA_LENGTH
             if add_row:
                 sample_length += 1
-            self.sample_table = _get_sample_table(
+            self.sample_table = get_sample_table(
                 table_format=self.data_format, n_rows=sample_length, add_column=add_column, change_types=change_type
             )
 
@@ -400,7 +302,7 @@ class TestComplexFormats(unittest.TestCase):
     expect_extra_column_fails: bool
 
     def setUp(self) -> None:
-        self.sample_table = _get_sample_table(table_format=self.data_format, n_rows=SAMPLE_DATA_LENGTH)
+        self.sample_table = get_sample_table(table_format=self.data_format, n_rows=SAMPLE_DATA_LENGTH)
         self.expected_parquet_digest = _get_table_digest(self.sample_table)
         self.data_dir = CACHE.get_location(ENVIRONMENT, self.__class__.__name__)
         self.parquet_file = self.data_dir / self.expected_parquet_digest
@@ -422,7 +324,7 @@ class TestComplexFormats(unittest.TestCase):
 
     def test_save_table_extra_column_fails(self) -> None:
         self.assertFalse(self.parquet_file.is_file())
-        sample_table = _get_sample_table(self.data_format, n_rows=SAMPLE_DATA_LENGTH, add_column=True)
+        sample_table = get_sample_table(self.data_format, n_rows=SAMPLE_DATA_LENGTH, add_column=True)
 
         if self.expect_extra_column_fails:
             with self.assertRaises(TableFormatError):
@@ -435,7 +337,7 @@ class TestComplexFormats(unittest.TestCase):
 
     def test_save_table_different_column_types_fails(self) -> None:
         self.assertFalse(self.parquet_file.is_file())
-        sample_table = _get_sample_table(self.data_format, n_rows=SAMPLE_DATA_LENGTH, change_types=True)
+        sample_table = get_sample_table(self.data_format, n_rows=SAMPLE_DATA_LENGTH, change_types=True)
 
         known_format = get_known_format(sample_table)
         with self.assertRaises(TableFormatError):
